fix: audit cleanup — DRY, bug fix, magic numbers, include hygiene
- Unify HeuristicRuleResult/RuleMatch (DRY): make HeuristicRuleResult a
  type alias for RuleMatch; simplify HybridDetectionService conversion
- Fix severity truncation bug in GrpcStreamSink: proto field changed from
  uint32 to float (was casting float [0.0,1.0] to uint32, always yielding 0)
- Add [[nodiscard]] on sweepExpiredFlows(); suppress internal callers with
  std::ignore
- Replace magic numbers: kStreamEventQueueCapacity, kShutdownPollInterval,
  kMinStdDevEpsilon; remove redundant maxConcurrentSessions override
- Replace std::ostringstream with std::format (C++23) in NidsServiceImpl
- Remove unused #include <memory> from HybridDetectionService.h
- ServiceRegistry: std::set -> std::unordered_set (key order not needed)
- Add // NOSONAR annotations to FilterPanel.cpp Qt allocations
- ServiceDialog: sort list widget items for display (unordered_set has no
  inherent ordering)

430 tests pass, zero warnings.

+++ src/server/GrpcStreamSink.cpp
@@ -71,7 +71,7 @@ void GrpcStreamSink::populateDetectionEvent(
         auto* match = event.add_rule_matches();
         match->set_rule_name(rule.ruleName);
         match->set_description(rule.description);
-        match->set_severity(static_cast<std::uint32_t>(rule.severity));
+        match->set_severity(rule.severity);
     }
 
     // Combined result
